migrate with-components test
