+++ tests/test-utils/helpers.js
@@ -4,6 +4,7 @@ const path = require('path');
 const exec = util.promisify(require('child_process').exec);
 
 const cheerio = require('cheerio');
+const { execRegexOnAll } = require('../../src/utils/general-helpers');
 
 /**
  * Builds the example
@@ -38,6 +39,32 @@ function getSelector(outPath) {
 }
 
 /**
+ * Transforms hash like snkajda to 1, 2, 3 like numbers
+ * @param {string} content style content
+ * @return {string}
+ */
+function transformHash(content) {
+  let newContent = '';
+  let currentNum = 0;
+  let lastIndex = 0;
+  const hashes = {};
+
+  const { matches } = execRegexOnAll(/\[data-abell-(.*?)\]/g, content);
+  for (const match of matches) {
+    if (!hashes[match[1]]) {
+      hashes[match[1]] = ++currentNum;
+    }
+    // prettier-ignore
+    // eslint-disable-next-line max-len
+    newContent += `${content.slice(lastIndex, match.index)}[data-abell-${currentNum}]`;
+    lastIndex = match.index + match[0].length;
+  }
+
+  newContent += content.slice(lastIndex);
+  return newContent.trim().replace(/\r/g, '');
+}
+
+/**
  * Returns responsive path that remains same even in windows
  * @param {string} pathString path to the file/folder
  * @return {string}
@@ -47,5 +74,6 @@ const resPath = (pathString) => pathString.replace(/\//g, path.sep);
 module.exports = {
   resPath,
   preTestSetup,
+  transformHash,
   getSelector
 };
